requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
fpdf==1.7.2
lxml==6.0.2
packaging==25.0
python-docx==1.2.0
typing_extensions==4.15.0
